Move factories etc. into IntegrationTestBase
The various factories are going to be required in different test classes, so makes sense to just define & setup in the base test class
